Fix for #227: Wrong grammatical case in german locale when formatting a list of durations
Fix for the logic that adjusts the german grammatical case for some units in
the `decorate` and `decorateUnrounded` methods.
The new logic replaces all occurrences of words that need to be adjusted in the
translated time string, the old logic did this only for single durations.

This should also fix issue #220, because the new logic simply replaces plural words
and doesn't check if the duration is a plural duration.

Contains also a minor improvement in SimpleTimeFormat.isPlural() that prevents
computing the same number twice.